8244737: Shenandoah: move mode code to gc/shenandoah/mode Reviewed-by: rkennke
