+++ src/hotspot/share/gc/shenandoah/mode/shenandoahIUMode.hpp
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019, Red Hat, Inc. All rights reserved.
+ * Copyright (c) 2020, Red Hat, Inc. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 only, as
@@ -21,20 +22,20 @@
  *
  */
 
-#ifndef SHARE_GC_SHENANDOAH_SHENANDOAHNORMALMODE_HPP
-#define SHARE_GC_SHENANDOAH_SHENANDOAHNORMALMODE_HPP
+#ifndef SHARE_GC_SHENANDOAH_MODE_SHENANDOAHIUMODE_HPP
+#define SHARE_GC_SHENANDOAH_MODE_SHENANDOAHIUMODE_HPP
 
-#include "gc/shenandoah/shenandoahMode.hpp"
+#include "gc/shenandoah/mode/shenandoahNormalMode.hpp"
 
 class ShenandoahHeuristics;
 
-class ShenandoahNormalMode : public ShenandoahMode {
+class ShenandoahIUMode : public ShenandoahNormalMode {
 public:
   virtual void initialize_flags() const;
-  virtual ShenandoahHeuristics* initialize_heuristics() const;
-  virtual const char* name()     { return "Normal"; }
+
+  virtual const char* name()     { return "Incremental-Update"; }
   virtual bool is_diagnostic()   { return false; }
-  virtual bool is_experimental() { return false; }
+  virtual bool is_experimental() { return true; }
 };
 
-#endif // SHARE_GC_SHENANDOAH_SHENANDOAHNORMALMODE_HPP
+#endif // SHARE_GC_SHENANDOAH_MODE_SHENANDOAHIUMODE_HPP
