+++ src/hotspot/share/gc/shenandoah/mode/shenandoahMode.hpp
@@ -21,8 +21,8 @@
  *
  */
 
-#ifndef SHARE_GC_SHENANDOAH_SHENANDOAHMODE_HPP
-#define SHARE_GC_SHENANDOAH_SHENANDOAHMODE_HPP
+#ifndef SHARE_GC_SHENANDOAH_MODE_SHENANDOAHMODE_HPP
+#define SHARE_GC_SHENANDOAH_MODE_SHENANDOAHMODE_HPP
 
 #include "memory/allocation.hpp"
 
@@ -53,4 +53,4 @@ public:
   virtual bool is_experimental() = 0;
 };
 
-#endif // SHARE_GC_SHENANDOAH_SHENANDOAHMODE_HPP
+#endif // SHARE_GC_SHENANDOAH_MODE_SHENANDOAHMODE_HPP
