+++ src/hotspot/share/gc/shenandoah/mode/shenandoahNormalMode.cpp
@@ -22,11 +22,11 @@
 
 #include "precompiled.hpp"
 #include "gc/shenandoah/shenandoahConcurrentRoots.hpp"
-#include "gc/shenandoah/shenandoahNormalMode.hpp"
 #include "gc/shenandoah/heuristics/shenandoahAdaptiveHeuristics.hpp"
 #include "gc/shenandoah/heuristics/shenandoahAggressiveHeuristics.hpp"
 #include "gc/shenandoah/heuristics/shenandoahCompactHeuristics.hpp"
 #include "gc/shenandoah/heuristics/shenandoahStaticHeuristics.hpp"
+#include "gc/shenandoah/mode/shenandoahNormalMode.hpp"
 #include "logging/log.hpp"
 #include "logging/logTag.hpp"
